+++ packages/grid/src/utils/clipboard/paste.ts
@@ -4,6 +4,9 @@ import {
     AITableField,
     AITableFieldType,
     AITableRecord,
+    createDefaultField,
+    createDefaultFieldName,
+    FieldOptions,
     FieldValue,
     getFieldValue,
     idCreator,
@@ -13,6 +16,7 @@ import {
 import { readFromClipboard, aiTableFragmentAttribute, extractText } from '../clipboard';
 import { FieldModelMap } from '../field/model';
 import { processPastedValueForSelect } from '../field/model/select';
+import { AddRecordOptions, AddFieldOptions } from '../../core';
 
 const aiTableAttributePattern = new RegExp(`${aiTableFragmentAttribute}="(.+?)"`, 'm');
 
@@ -122,22 +126,27 @@ function getPasteValue(
     let originData = field && record ? { field, cellValue: getFieldValue(record, field) } : null;
     if (targetField.type === AITableFieldType.select) {
         let { existOptionIds, newOptions } = processPastedValueForSelect(plainText, targetField, originData);
+        let newField: AITableField | null = null;
+        let newOptionIds: string[] = [];
 
-        newOptions = newOptions.map((option) => {
-            return {
-                ...option,
-                _id: idCreator()
+        if (newOptions.length) {
+            newOptions = newOptions.map((option) => {
+                return {
+                    ...option,
+                    _id: idCreator()
+                };
+            });
+            newField = {
+                ...targetField,
+                settings: {
+                    ...targetField.settings,
+                    options: [...((targetField.settings as SelectSettings)?.options || []), ...newOptions]
+                }
             };
-        });
-        const newField = {
-            ...targetField,
-            settings: {
-                ...targetField.settings,
-                options: [...((targetField.settings as SelectSettings)?.options || []), ...newOptions]
-            }
-        };
-        const newOptionIds = newOptions.map((option) => option._id).filter((id) => !!id) as string[];
-        const selectFieldValue = [...existOptionIds, ...newOptionIds];
+            newOptionIds = newOptions.map((option) => option._id).filter((id) => !!id) as string[];
+        }
+
+        const selectFieldValue = newOptionIds?.length ? [...existOptionIds, ...newOptionIds] : existOptionIds;
         return {
             value: selectFieldValue,
             newField
@@ -150,6 +159,35 @@ function getPasteValue(
 export interface AITablePasteActions {
     updateFieldValue: (data: UpdateFieldValueOptions) => void;
     setField: (field: AITableField) => void;
+    addRecord: (data: AddRecordOptions) => void;
+    addField: (data: AddFieldOptions) => void;
+}
+
+function appendRecord(aiTable: AITable, actions: AITablePasteActions) {
+    const allRecords = aiTable.records();
+    const lastRecordId = allRecords.length > 0 ? allRecords[allRecords.length - 1]._id : '';
+    actions.addRecord({
+        originId: lastRecordId
+    });
+}
+
+function appendField(aiTable: AITable, originField: AITableField | null, actions: AITablePasteActions) {
+    const lastFieldId = aiTable.fields().length > 0 ? aiTable.fields()[aiTable.fields().length - 1]._id : '';
+    let defaultFieldValue: Partial<AITableField>;
+    if (originField) {
+        defaultFieldValue = {
+            ...originField,
+            name: createDefaultFieldName(aiTable, FieldOptions.find((item) => item.type === originField.type)!),
+            _id: idCreator()
+        };
+    } else {
+        defaultFieldValue = createDefaultField(aiTable, AITableFieldType.text);
+    }
+
+    actions.addField({
+        originId: lastFieldId,
+        defaultValue: defaultFieldValue
+    });
 }
 
 export const writeToAITable = async (aiTable: AITable, actions: AITablePasteActions) => {
@@ -166,20 +204,24 @@ export const writeToAITable = async (aiTable: AITable, actions: AITablePasteActi
     const [startRecordId, startFieldId] = firstCell.split(':');
     const startRowIndex = aiTable.context!.visibleRowsIndexMap().get(startRecordId) ?? 0;
     const startColIndex = aiTable.context!.visibleColumnsIndexMap().get(startFieldId) ?? 0;
-    const visibleFields = AITable.getVisibleFields(aiTable);
-    const linearRows = aiTable.context!.linearRows();
     const references = aiTable.context!.references();
+    let isPasteSuccess = false;
 
     clipboardContent.forEach((row, i) => {
+        const targetRowIndex = startRowIndex + i;
+        if (targetRowIndex >= aiTable.context!.linearRows().length - 1) {
+            appendRecord(aiTable, actions);
+        }
+
         row.forEach((plainText, j) => {
-            const targetRowIndex = startRowIndex + i;
             const targetColIndex = startColIndex + j;
-            if (targetRowIndex >= linearRows.length || targetColIndex >= visibleFields.length) {
-                return;
+            if (targetColIndex >= AITable.getVisibleFields(aiTable).length) {
+                const originField = aiTableContent?.fields[j] || null;
+                appendField(aiTable, originField, actions);
             }
 
-            const targetRecord = linearRows[targetRowIndex];
-            const targetField = visibleFields[targetColIndex];
+            const targetRecord = aiTable.context!.linearRows()[targetRowIndex];
+            const targetField = AITable.getVisibleFields(aiTable)[targetColIndex];
             const recordIndex = i;
             const fieldIndex = j;
             const { value, newField } = getPasteValue(plainText, aiTableContent, recordIndex, fieldIndex, targetField, references);
@@ -189,11 +231,16 @@ export const writeToAITable = async (aiTable: AITable, actions: AITablePasteActi
             }
 
             if (value !== null) {
-                actions.updateFieldValue({
-                    value,
-                    path: [targetRecord._id, targetField._id]
-                });
+                try {
+                    actions.updateFieldValue({
+                        value,
+                        path: [targetRecord._id, targetField._id]
+                    });
+                    isPasteSuccess = true;
+                } catch (error) {}
             }
         });
     });
+
+    return isPasteSuccess;
 };